refactor(auth,core): delegate GitHub validation to github-core service
