+++ github-core/src/main/java/com/itachallenge/githubcore/service/GithubOAuthService.java
@@ -1,10 +1,9 @@
-package com.itachallenge.auth.service;
+package com.itachallenge.githubcore.service;
 
 import reactor.core.publisher.Mono;
 import java.util.Map;
 
-public interface IAuthService {
+public interface GithubOAuthService {
     Mono<Map<String, Object>> validateTokenWithGithub(String token);
     Mono<String> exchangeCodeForToken(String code);
-
 }
